# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Federation API
  version: 0.1.0
  description: |
    Backend REST, ActivityPub, C2S, S2S inbox, WebFinger, and delivery administration contract.

    Bounded REST list operations return JSON arrays for backward compatibility and expose a uniform
    offset-pagination contract through `Link`, `X-Pagination-Limit`, `X-Pagination-Offset`,
    `X-Pagination-Count`, and `X-Pagination-Has-More` response headers. The `Link` header preserves
    active filters and supplies RFC 8288 `prev` and `next` relations when applicable.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/me/federation/inbox:
    get:
      tags: [Federation]
      operationId: listPersonalFederationInbox
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TicketPriority"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TicketType"
        - name: assignee
          in: query
          schema:
            type: string
            enum: [me, unassigned]
        - name: assignee_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Normalized ActivityPub activities delivered to the authenticated user's inbox.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FederationInboxActivity"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me/federation/discover:
    post:
      tags: [Federation]
      operationId: discoverPersonalFederationActor
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/RemoteFederationResource"
      responses:
        "200":
          description: Remote ActivityPub actor resolved and cached for federation actions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FederationRemoteActor"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/v1/me/federation/follows:
    get:
      tags: [Federation]
      operationId: listPersonalFederationFollows
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Remote actors followed by the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FederationRemoteFollow"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Federation]
      operationId: followPersonalFederationActor
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/RemoteFederationResource"
      responses:
        "200":
          description: The authenticated user already follows this remote actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowRemoteActorResult"
        "202":
          description: Follow activity created and queued for federation delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowRemoteActorResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/me/remote-project-invites:
    get:
      tags: [Federation]
      operationId: listRemoteProjectInvites
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Remote project invites addressed to the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteProjectInvite"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me/remote-project-invites/{inviteID}/accept:
    post:
      tags: [Federation]
      operationId: acceptRemoteProjectInvite
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InviteID"
      responses:
        "202":
          description: Remote project invite accepted and response delivery queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteProjectInviteResult"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/me/remote-project-invites/{inviteID}/reject:
    post:
      tags: [Federation]
      operationId: rejectRemoteProjectInvite
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InviteID"
      responses:
        "202":
          description: Remote project invite rejected and response delivery queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteProjectInviteResult"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/remote-projects:
    get:
      tags: [Federation]
      operationId: listRemoteProjects
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Accepted remote project workspaces visible to the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteProject"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/remote-projects/{remoteProjectID}:
    parameters:
      - $ref: "#/components/parameters/RemoteProjectID"
    get:
      tags: [Federation]
      operationId: getRemoteProject
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accepted remote project workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteProject"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/remote-projects/{remoteProjectID}/tickets:
    parameters:
      - $ref: "#/components/parameters/RemoteProjectID"
    get:
      tags: [Federation]
      operationId: listRemoteProjectTickets
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Remote project tickets fetched through signed ActivityPub reads.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteTicket"
        "404":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
    post:
      tags: [Federation]
      operationId: createRemoteTicket
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/CreateTicket"
      responses:
        "202":
          description: Remote ticket Create activity queued for delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteTicketWriteResult"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/remote-projects/{remoteProjectID}/tickets/{remoteTicketID}:
    parameters:
      - $ref: "#/components/parameters/RemoteProjectID"
      - $ref: "#/components/parameters/RemoteTicketID"
    get:
      tags: [Federation]
      operationId: getRemoteTicket
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Remote ticket fetched through a signed ActivityPub read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteTicket"
        "404":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Federation]
      operationId: updateRemoteTicket
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/UpdateTicket"
      responses:
        "202":
          description: Remote ticket Update activity queued for delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteTicketWriteResult"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Federation]
      operationId: deleteRemoteTicket
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Remote ticket Delete activity queued for delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteTicketWriteResult"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/remote-projects/{remoteProjectID}/tickets/{remoteTicketID}/move:
    parameters:
      - $ref: "#/components/parameters/RemoteProjectID"
      - $ref: "#/components/parameters/RemoteTicketID"
    post:
      tags: [Federation]
      operationId: moveRemoteTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/TicketStatus"
      responses:
        "202":
          description: Remote ticket status Update activity queued for delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteTicketWriteResult"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: progo.session
      description: HttpOnly SameSite=Strict browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsBearerAuth:
      type: http
      scheme: bearer
      description: Optional in development; required when METRICS_TOKEN is configured.
...
  parameters:
...
    InviteID:
      name: inviteID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RemoteProjectID:
      name: remoteProjectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RemoteTicketID:
      name: remoteTicketID
      in: path
      required: true
      description: URL-safe base64 encoding of the remote ticket ActivityPub ID.
      schema:
        type: string
...
  requestBodies:
...
    RemoteFederationResource:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [resource]
            properties:
              resource:
                type: string
                description: "Remote acct: resource, bare handle, or actor URL."
              target:
                type: string
                deprecated: true
...
    CreateTicket:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [title, priority, type]
            properties:
              title:
                type: string
              description:
                type: string
              priority:
                $ref: "#/components/schemas/TicketPriority"
              type:
                $ref: "#/components/schemas/TicketType"
              parent_id:
                type:
                  - string
                  - "null"
                format: uuid
              assignee_id:
                type:
                  - string
                  - "null"
                format: uuid
              due_date:
                type: string
                format: date
              label_ids:
                type: array
                maxItems: 20
                uniqueItems: true
                items:
                  type: string
                  format: uuid
...
  responses:
...
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    FederationRemoteActor:
      type: object
      required: [id, ap_id, type, preferred_username, handle, name, summary, inbox_url, outbox_url, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        ap_id:
          type: string
          format: uri
        type:
          type: string
        preferred_username:
          type: string
        handle:
          type: string
        name:
          type: string
        summary:
          type: string
        inbox_url:
          type: string
          format: uri
        outbox_url:
          type: string
          format: uri
        followers_url:
          type: string
          format: uri
        following_url:
          type: string
          format: uri
        last_fetched_at:
          type: string
          format: date-time
        fetch_error:
          type: string
        fetch_error_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FederationInboxActivity:
      type: object
      required:
        - id
        - activity_ap_id
        - activity_type
        - actor_id
        - actor_ap_id
        - actor_type
        - actor_handle
        - actor_name
        - received_at
        - created_at
      properties:
        id:
          type: string
          format: uuid
        activity_ap_id:
          type: string
          format: uri
        activity_type:
          type: string
        actor_id:
          type: string
          format: uuid
        actor_ap_id:
          type: string
          format: uri
        actor_type:
          type: string
        actor_handle:
          type: string
        actor_name:
          type: string
        object_ap_id:
          type: string
          format: uri
        object_type:
          type: string
        object_name:
          type: string
        object_content:
          type: string
        target_ap_id:
          type: string
          format: uri
        target_actor_id:
          type: string
          format: uuid
        target_type:
          type: string
        target_handle:
          type: string
        target_name:
          type: string
        received_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    FederationRemoteFollow:
      type: object
      required:
        - actor_id
        - actor_ap_id
        - actor_type
        - preferred_username
        - handle
        - name
        - summary
        - inbox_url
        - outbox_url
        - state
        - created_at
        - updated_at
      properties:
        actor_id:
          type: string
          format: uuid
        actor_ap_id:
          type: string
          format: uri
        actor_type:
          type: string
        preferred_username:
          type: string
        handle:
          type: string
        name:
          type: string
        summary:
          type: string
        inbox_url:
          type: string
          format: uri
        outbox_url:
          type: string
          format: uri
        followers_url:
          type: string
          format: uri
        following_url:
          type: string
          format: uri
        state:
          type: string
          enum: [pending, accepted, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FollowRemoteActorResult:
      type: object
      required: [follow, created]
      properties:
        follow:
          $ref: "#/components/schemas/FederationRemoteFollow"
        delivery:
          $ref: "#/components/schemas/FederationFollowDelivery"
        created:
          type: boolean
    RemoteProjectInvite:
      type: object
      required:
        - id
        - invite_ap_id
        - activity_id
        - project_ap_id
        - project_name
        - inviter_actor_id
        - inviter_ap_id
        - inviter_handle
        - inviter_name
        - invitee_actor_id
        - role
        - role_permissions
        - target_inbox_url
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        invite_ap_id:
          type: string
          format: uri
        activity_id:
          type: string
          format: uuid
        project_ap_id:
          type: string
          format: uri
        project_name:
          type: string
        inviter_actor_id:
          type: string
          format: uuid
        inviter_ap_id:
          type: string
          format: uri
        inviter_handle:
          type: string
        inviter_name:
          type: string
        invitee_actor_id:
          type: string
          format: uuid
        role:
          type: string
        role_permissions:
          type: array
          items:
            type: string
        target_inbox_url:
          type: string
          format: uri
        status:
          type: string
          enum: [pending, accepted, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
    RemoteProjectInviteResult:
      type: object
      required: [invite]
      properties:
        invite:
          $ref: "#/components/schemas/RemoteProjectInvite"
        delivery:
          $ref: "#/components/schemas/FederationFollowDelivery"
    RemoteProject:
      type: object
      required:
        - id
        - project_ap_id
        - project_name
        - role
        - role_permissions
        - target_inbox_url
        - inviter_actor_id
        - inviter_ap_id
        - inviter_handle
        - inviter_name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_ap_id:
          type: string
          format: uri
        project_name:
          type: string
        role:
          type: string
        role_permissions:
          type: array
          items:
            type: string
        target_inbox_url:
          type: string
          format: uri
        inviter_actor_id:
          type: string
          format: uuid
        inviter_ap_id:
          type: string
          format: uri
        inviter_handle:
          type: string
        inviter_name:
          type: string
        remote_actor_id:
          type:
            - string
            - "null"
          format: uuid
        remote_handle:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type:
            - string
            - "null"
          format: date-time
    RemoteTicket:
      type: object
      required: [id, ap_id, title, description, status, priority, type, project_id, reporter_id, is_resolved, created_at, updated_at]
      properties:
        id:
          type: string
          description: URL-safe base64 encoding of the ticket ActivityPub ID.
        ap_id:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        type:
          $ref: "#/components/schemas/TicketType"
        rank:
          type: string
        parent_id:
          type:
            - string
            - "null"
        project_id:
          type: string
          format: uuid
        reporter_id:
          type: string
        assignee_id:
          type:
            - string
            - "null"
        is_resolved:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RemoteTicketWriteResult:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/RemoteTicket"
        delivery:
          $ref: "#/components/schemas/FederationFollowDelivery"
...
    TicketPriority:
      type: string
      enum: [low, medium, high, urgent]
    TicketStatus:
      type: string
      enum: [open, in_progress, review, done]
    TicketType:
      type: string
      enum: [epic, task, subtask]
...